+++ example.js
@@ -1,5 +1,5 @@
+const fs = require('fs')
 const MBTiles = require('.')
-const schema = require('./schema')
 
 const mbtiles = new MBTiles('./fixtures/empty.mbtiles')
 // mb.metadata(data => console.log(data))
@@ -11,13 +11,16 @@ const options = {
   format: 'png',
   bounds: [-110, -40, 95, 50]
 }
+const image = fs.readFileSync('./fixtures/images/0/0/0.png')
 
-mbtiles.metadata()
-  .then(metadata => console.log(metadata))
-mbtiles.count()
-  .then(count => console.log(count))
-mbtiles.update(options)
-  .then(metadata => console.log(metadata))
-mbtiles.tables()
+// mbtiles.metadata()
+//   .then(metadata => console.log(metadata))
+// mbtiles.count()
+//   .then(count => console.log(count))
+// mbtiles.update(options)
+//   .then(metadata => console.log(metadata))
+// mbtiles.tables()
+//   .then(() => console.log('done'))
+
+mbtiles.save([0, 0, 0], image)
   .then(() => console.log('done'))
-